xtensa/esp32: Changes in rtc driver to support rtc wdt driver
